+++ src/SexyAppFramework/Buffer.h
@@ -7,7 +7,7 @@
 namespace Sexy
 {
 
-typedef std::vector<uchar> ByteVector;
+typedef std::vector<uint8_t> ByteVector;
 
 class Buffer
 {
@@ -25,7 +25,7 @@ class Buffer
 	void Clear();
 
 	void FromWebString(const std::string &theString);
-	void WriteByte(uchar theByte);
+	void WriteByte(uint8_t theByte);
 	void WriteNumBits(int theNum, int theBits);
 	static int GetBitsRequired(int theNum, bool isSigned);
 	void WriteBoolean(bool theBool);
@@ -35,13 +35,13 @@ class Buffer
 	void WriteUTF8String(const std::wstring &theString);
 	void WriteLine(const std::string &theString);
 	void WriteBuffer(const ByteVector &theBuffer);
-	void WriteBytes(const uchar *theByte, int theCount);
+	void WriteBytes(const uint8_t *theByte, int theCount);
 	void SetData(const ByteVector &theBuffer);
-	void SetData(uchar *thePtr, int theCount);
+	void SetData(uint8_t *thePtr, int theCount);
 
 	std::string ToWebString() const;
 	std::wstring UTF8ToWideString() const;
-	uchar ReadByte() const;
+	uint8_t ReadByte() const;
 	int ReadNumBits(int theBits, bool isSigned) const;
 	bool ReadBoolean() const;
 	short ReadShort() const;
@@ -49,13 +49,13 @@ class Buffer
 	std::string ReadString() const;
 	std::wstring ReadUTF8String() const;
 	std::string ReadLine() const;
-	void ReadBytes(uchar *theData, int theLen) const;
+	void ReadBytes(uint8_t *theData, int theLen) const;
 	void ReadBuffer(ByteVector *theByteVector) const;
 
-	const uchar *GetDataPtr() const;
+	const uint8_t *GetDataPtr() const;
 	int GetDataLen() const;
 	int GetDataLenBits() const;
-	ulong GetCRC32(ulong theSeed = 0) const;
+	uint32_t GetCRC32(uint32_t theSeed = 0) const;
 
 	bool AtEnd() const;
 	bool PastEnd() const;
